% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{KSN_biogrid_yeast}
\alias{KSN_biogrid_yeast}
\title{The KSN of biogrid for yeast to run Kinase activity estimation}
\format{
The full KSN contains 15553 Kinase substrate interaction for 146 unique kinases
\describe{
    \item{psite}{substrate identifier with gene name, residue and position}
    \item{kinase}{kinase name}
}
}
\source{
\url{https://github.com/saezlab/conformationomic_yeast_picotti_2020/blob/master/scripts/generate_KSN.R}
}
\description{
Using common yeast gene names
}
\examples{
data("KSN_biogrid_yeast")
}
\keyword{KSN}
