% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PKN_STRING_yeast}
\alias{PKN_STRING_yeast}
\title{The PKN of yeast from STRING}
\format{
The full PKN contains 4662 signed directed protein protein interactions
\describe{
    \item{source}{upstream protein}
    \item{targets}{downstream protein}
    \item{sign}{activation/inhibition (1/-1)}
}
}
\source{
\url{https://github.com/saezlab/conformationomic_yeast_picotti_2020/blob/master/scripts/format_STRING_PKN_signed.R}
}
\description{
Filtered interactions comming from database only, with 700 confidence score at least
}
\examples{
data("PKN_STRING_yeast")
}
\keyword{PKN}
