% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TFTN_compound_yeast}
\alias{TFTN_compound_yeast}
\title{The TF-target network of yeast from JASPAR motif recognition and yeastract curated database}
\format{
The full TFTN contains 14183 signed interactions
\describe{
    \item{tf}{upstream transcription factor}
    \item{target}{downstream transcript}
    \item{mor}{activation/inhibition of transcription(1/-1)}
}
}
\source{
\url{https://github.com/saezlab/yeast_phospho/blob/master/files/tf_gene_network_binding_sites_posterior_90.tab}

\url{https://github.com/ptgm/Paper_2020_Sci_Rep_scripts/blob/master/yeast2019-full-conds-net.txt}
}
\description{
The sign of each TF-target network interaction is determined from a large collection of gene KO experiement of TFs
to determine the score of the jaspar interactions, the sign of the "value" column of the original table of jaspar (see source) is multiplied by -1
}
\examples{
data("TFTN_compound_yeast")
}
\keyword{TFTN}
