% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TFTN_JASPAR_yeast}
\alias{TFTN_JASPAR_yeast}
\title{The TF-target network of yeast from JASPAR motif recognition}
\format{
The full TFTN contains 4201 signed interactions
\describe{
    \item{tf}{upstream transcription factor}
    \item{target}{downstream transcript}
    \item{mor}{activation/inhibition of transcription(1/-1)}
}
}
\source{
\url{https://github.com/saezlab/yeast_phospho/blob/master/files/tf_gene_network_binding_sites_posterior_90.tab}
}
\description{
The sign of each TF-target network interaction is determined from a large collection of gene KO experiement of TFs
to determine the score, the sign of the "value" column of the original table (see source) is multiplied by -1
}
\examples{
data("TFTN_JASPAR_yeast")
}
\keyword{TFTN}
